refactor: move Ion parsers to literal/ion/ subdir
ion_null.rs → ion/null.rs
ion_number.rs → ion/number.rs
ion_string.rs → ion/string.rs
ion_timestamp.rs → ion/timestamp.rs
annotation.rs → ion/annotation.rs
