+++ partiql-winnow-parser/src/literal/null_strategy.rs
@@ -1,6 +1,6 @@
 //! NullMissingStrategy — `null`, `null.int`, `MISSING`.
 
-use super::ion_null;
+use super::ion::null;
 use super::LiteralStrategy;
 use crate::expr::StrategyContext;
 use partiql_ast::ast;
@@ -11,10 +11,10 @@ pub struct NullMissingStrategy;
 
 impl LiteralStrategy for NullMissingStrategy {
     fn parse<'a>(&self, input: &mut &'a str, ctx: &StrategyContext<'_>) -> PResult<ast::Expr> {
-        if ion_null::missing(input).is_ok() {
+        if null::missing(input).is_ok() {
             return Ok(ast::Expr::Lit(ctx.node(Lit::Missing)));
         }
-        let _ = ion_null::ion_null(input)?;
+        let _ = null::ion_null(input)?;
         Ok(ast::Expr::Lit(ctx.node(Lit::Null)))
     }
