+++ partiql-winnow-parser/src/literal/number_strategy.rs
@@ -1,6 +1,6 @@
 //! NumericLiteralStrategy — integer, decimal, float.
 
-use super::ion_number;
+use super::ion::number;
 use super::LiteralStrategy;
 use crate::expr::StrategyContext;
 use partiql_ast::ast;
@@ -11,11 +11,11 @@ pub struct NumericLiteralStrategy;
 
 impl LiteralStrategy for NumericLiteralStrategy {
     fn parse<'a>(&self, input: &mut &'a str, ctx: &StrategyContext<'_>) -> PResult<ast::Expr> {
-        let num = ion_number::ion_number(input)?;
+        let num = number::ion_number(input)?;
         Ok(match num {
-            ion_number::IonNumber::Integer(n) => ast::Expr::Lit(ctx.node(Lit::Int64Lit(n))),
-            ion_number::IonNumber::Decimal(d) => ast::Expr::Lit(ctx.node(Lit::DecimalLit(d))),
-            ion_number::IonNumber::Float(f) => ast::Expr::Lit(ctx.node(Lit::DoubleLit(f))),
+            number::IonNumber::Integer(n) => ast::Expr::Lit(ctx.node(Lit::Int64Lit(n))),
+            number::IonNumber::Decimal(d) => ast::Expr::Lit(ctx.node(Lit::DecimalLit(d))),
+            number::IonNumber::Float(f) => ast::Expr::Lit(ctx.node(Lit::DoubleLit(f))),
         })
     }
